Arumando folders corretamente
