+++ lib/app/modules/materias/models/materias_model.dart
@@ -1,4 +1,6 @@
 import 'package:system_maua_front/app/modules/materias/models/professores_model.dart';
+import 'package:system_maua_front/app/modules/materias/pages/materias-apresentacao/models/materias_apresentacao_model.dart';
+import 'package:system_maua_front/app/modules/materias/pages/materias-aulas/models/materias_aulas_model.dart';
 import 'package:system_maua_front/app/shared/enumerates/tipo_materia_enum.dart';
 
 class MateriasModel {
@@ -6,8 +8,12 @@ class MateriasModel {
   final TipoMateriaEnum? nomeMateria;
   final String? foto;
   final List<ProfessoresModel>? professores;
+  final List<MateriasAulasModel>? listaAulas;
+  final MateriasApresentacaoModel? apresentacaoModel;
 
   MateriasModel({
+    this.apresentacaoModel,
+    this.listaAulas,
     required this.codigoMateria,
     required this.nomeMateria,
     this.foto,
@@ -16,6 +22,12 @@ class MateriasModel {
 
   factory MateriasModel.newInstance() {
     return MateriasModel(
-        codigoMateria: '', foto: '', nomeMateria: null, professores: null);
+      apresentacaoModel: null,
+      codigoMateria: '',
+      foto: '',
+      nomeMateria: null,
+      professores: null,
+      listaAulas: null,
+    );
   }
 }
