+++ lib/app/modules/materias/pages/file-view/file_view_page.dart
@@ -2,8 +2,9 @@ import 'package:advance_pdf_viewer/advance_pdf_viewer.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:flutter_modular/flutter_modular.dart';
-import 'package:system_maua_front/app/modules/file-view/file_view_controller.dart';
 import 'package:system_maua_front/app/shared/components/app_bar/app_bar_widget.dart';
+
+import 'file_view_controller.dart';
 
 class FileViewPage extends StatefulWidget {
   FileViewPage({Key? key}) : super(key: key);
