+++ lib/app/modules/materias/pages/materias-apresentacao/materias_apresentacao_controller.dart
@@ -1,5 +1,6 @@
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:mobx/mobx.dart';
+import 'package:system_maua_front/app/modules/materias/models/professores_model.dart';
 import 'package:system_maua_front/app/modules/materias/repositories/materias_repository_interface.dart';
 
 import 'models/materias_apresentacao_model.dart';
@@ -21,9 +22,13 @@ abstract class _MateriasApresentacaoControllerBase with Store {
   @observable
   MateriasApresentacaoModel materia = MateriasApresentacaoModel.newInstance();
 
+  @observable
+  List<ProfessoresModel> professores = [];
+
   @action
   Future<void> getMateria() async {
-    materia = await repository.getMateriaEspecifica(codigoMateria);
+    materia = await repository.getApresentacao(codigoMateria);
+    professores = await repository.getProfessores(codigoMateria);
   }
 
   void navigateToPlanoEnsino() async {
