+++ lib/app/modules/materias/pages/materias-aulas/materias_aulas_page.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_modular/flutter_modular.dart';
-import 'package:system_maua_front/app/modules/materias-aulas/materias_aulas_controller.dart';
 import 'package:system_maua_front/app/shared/components/app_bar/app_bar_widget.dart';
+
+import 'materias_aulas_controller.dart';
 
 class MateriasAulasPage extends StatefulWidget {
   final String nomeMateria;
@@ -15,14 +16,16 @@ class _MateriasAulasPageState
     extends ModularState<MateriasAulasPage, MateriasAulasController> {
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBarWidget(
-        title: widget.nomeMateria,
-      ),
-      body: Container(
-        color: Colors.red,
-        height: 100,
-        width: 100,
+    return SafeArea(
+      child: Scaffold(
+        appBar: AppBarWidget(
+          title: widget.nomeMateria,
+        ),
+        body: Container(
+          color: Colors.red,
+          height: 100,
+          width: 100,
+        ),
       ),
     );
   }
