+++ lib/app/modules/materias/pages/modal-buttom-sheet/widgets/type_ahead_field_widget.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_typeahead/flutter_typeahead.dart';
-import 'package:system_maua_front/app/modules/modal-buttom-sheet/models/generic_fields_model.dart';
+import 'package:system_maua_front/app/modules/materias/pages/modal-buttom-sheet/models/generic_fields_model.dart';
 
 class TypeAheadFieldWidget extends StatelessWidget {
   final String? titulo;
