+++ test/app/modules/materias/materias_controller_test.mocks.dart
@@ -5,12 +5,14 @@
 import 'dart:async' as _i5;
 
 import 'package:mockito/mockito.dart' as _i1;
-import 'package:system_maua_front/app/modules/materias-apresentacao/models/materias_apresentacao_model.dart'
-    as _i2;
-import 'package:system_maua_front/app/modules/materias-aulas/models/materias_aulas_model.dart'
-    as _i3;
 import 'package:system_maua_front/app/modules/materias/models/materias_model.dart'
     as _i6;
+import 'package:system_maua_front/app/modules/materias/models/professores_model.dart'
+    as _i7;
+import 'package:system_maua_front/app/modules/materias/pages/materias-apresentacao/models/materias_apresentacao_model.dart'
+    as _i2;
+import 'package:system_maua_front/app/modules/materias/pages/materias-aulas/models/materias_aulas_model.dart'
+    as _i3;
 import 'package:system_maua_front/app/modules/materias/repositories/materias_repository_interface.dart'
     as _i4;
 
@@ -42,13 +44,19 @@ class MockIMateriasRepository extends _i1.Mock
                   Future<List<_i6.MateriasModel>>.value(<_i6.MateriasModel>[]))
           as _i5.Future<List<_i6.MateriasModel>>);
   @override
-  _i5.Future<_i2.MateriasApresentacaoModel> getMateriaEspecifica(
+  _i5.Future<_i2.MateriasApresentacaoModel> getApresentacao(
           String? codigoMateria) =>
-      (super.noSuchMethod(
-              Invocation.method(#getMateriaEspecifica, [codigoMateria]),
+      (super.noSuchMethod(Invocation.method(#getApresentacao, [codigoMateria]),
               returnValue: Future<_i2.MateriasApresentacaoModel>.value(
                   _FakeMateriasApresentacaoModel()))
           as _i5.Future<_i2.MateriasApresentacaoModel>);
+  @override
+  _i5.Future<List<_i7.ProfessoresModel>> getProfessores(
+          String? codigoMateria) =>
+      (super.noSuchMethod(Invocation.method(#getProfessores, [codigoMateria]),
+              returnValue: Future<List<_i7.ProfessoresModel>>.value(
+                  <_i7.ProfessoresModel>[]))
+          as _i5.Future<List<_i7.ProfessoresModel>>);
   @override
   _i5.Future<String> getUrlPdf(String? codigoMateria) =>
       (super.noSuchMethod(Invocation.method(#getUrlPdf, [codigoMateria]),
